+++ game/NPC_reactions.c
@@ -119,17 +119,12 @@ static void NPC_CheckAttacker( gentity_t *other, int mod ) {
 }
 
 void NPC_SetPainEvent( gentity_t *self ) {
-	if ( !self->NPC || !(self->NPC->aiFlags&NPCAI_DIE_ON_IMPACT) ) {
-		// no more borg
-		//	if( self->client->playerTeam != TEAM_BORG )
-		//	{
-		//if ( !Q3_TaskIDPending( self, TID_CHAN_VOICE ) )
+	if ( !self->NPC || !(self->NPC->aiFlags & NPCAI_DIE_ON_IMPACT) ) {
 		if ( !trap->ICARUS_TaskIDPending( (sharedEntity_t *)self, TID_CHAN_VOICE ) && self->client ) {
-			//G_AddEvent( self, EV_PAIN, floor((float)self->health/self->max_health*100.0f) );
-			G_AddEvent( self, EV_PAIN, floor( (float)self->health / self->client->ps.stats[STAT_MAX_HEALTH] * 100.0f ) );
+			//G_AddEvent( self, EV_PAIN, floorf( (float)self->health / self->max_health * 100.0f ) );
+			G_AddEvent( self, EV_PAIN, floorf( (float)self->health / self->client->ps.stats[STAT_MAX_HEALTH] * 100.0f ) );
 			//rwwFIXMEFIXME: Do this properly?
 		}
-		//	}
 	}
 }
