+++ ui/ui_main.c
@@ -368,13 +368,13 @@ int UI_ParseAnimationFile( const char *filename, animation_t *animset, qboolean 
 			fps = 1;//Don't allow divide by zero error
 		}
 		if ( fps < 0 ) {//backwards
-			animset[animNum].frameLerp = floor( 1000.0f / fps );
+			animset[animNum].frameLerp = floorf( 1000.0f / fps );
 		}
 		else {
-			animset[animNum].frameLerp = ceil( 1000.0f / fps );
+			animset[animNum].frameLerp = ceilf( 1000.0f / fps );
 		}
 
-		//		animset[animNum].initialLerp = ceil(1000.0f / fabs(fps));
+		//animset[animNum].initialLerp = ceilf(1000.0f / fabsf(fps));
 	}
 
 #ifdef _DEBUG
@@ -2360,7 +2360,7 @@ static void UI_DrawServerRefreshDate( rectDef_t *rect, float scale, const vector
 		lowLight.g = 0.8f * color->g;
 		lowLight.b = 0.8f * color->b;
 		lowLight.a = 0.8f * color->a;
-		LerpColor( color, &lowLight, &newColor, 0.5f + 0.5f*sin( (float)(uiInfo.uiDC.realTime / PULSE_DIVISOR) ) );
+		LerpColor( color, &lowLight, &newColor, 0.5f + 0.5f*sinf( (float)(uiInfo.uiDC.realTime / PULSE_DIVISOR) ) );
 
 		trap->SE_GetStringTextString( "MP_INGAME_GETTINGINFOFORSERVERS", holdSPString, sizeof(holdSPString) );
 		Text_Paint( rect->x, rect->y, scale, &newColor, va( (char *)holdSPString, trap->LAN_GetServerCount( UI_SourceForLAN() ) ), 0, 0, textStyle, iMenuFont );
